Converted the old PhysVector (Vector2) from something silly to something useful
+ Most of these changes come from looking at game engines that handle
vectors as well as additional refreshing of my understanding of vectors
in physics and math
++ It's been a long time :/

+++ src/asteroids/Asteroids.java
@@ -17,7 +17,7 @@ public class Asteroids extends JPanel implements KeyListener {
         setBackground(Color.BLACK);
         setPreferredSize(new Dimension(500, 500));
 
-        player = new Ship((int) getPreferredSize().getWidth() / 2, (int) getPreferredSize().getHeight() / 2);
+        player = new Ship(getPreferredSize().getWidth() / 2, getPreferredSize().getHeight() / 2);
     }
 
     public void paint(Graphics g) {
